Plugins: port HTML plugin
