+++ avocado/core/resources/htmlresult/templates/report.mustache
